Document how jsbundling-rails compares to webpacker (#20)
With `webpacker` previously becoming the default JavaScript bundler in
Rails (https://weblog.rubyonrails.org/2019/8/15/Rails-6-0-final-release)
it's likely many folks are going to come to the `jsbundling-rails`
project wondering how the two compare. So let's document some of the
differences that will help developers decide if they want to migrate
to `jsbundling-rails` or stick with `webpacker`.