Updated the webpacker comparison
This _should be_ purely a formatting change, with no significant changes to content.  Github Flavored Markdown tables are annoying, and blending HTML tables with markdown isn't much better.  But I personally find the tables _so much_ easier to read than the list-item's flow of text.

Unfortunately, this does require horizontal scrolling or landscape orientation to be read on phone-sized screens.  It's not unusable, but it's not great either.  This _should_ be fixable with some CSS (style attributes, because GFM disables style tags).  But I want to know if this PR is broadly acceptable first, before making that effort. 🙂

Having the project names repeated at the beginning of each cell felt noisy to me, so I removed them from the start of sentences when the remaining sentence fragments read okay with them (to me, at least).   This removes context from the plain-text version that the rendered version retains by using columns.  So I added HTML comments so the plain-text markdown is still readable.  Additionally, some of the list items placed the `jsbundling` behavior first, and some placed the `webpacker`/`shakapacker` behavior first, so this normalizes that.