Test installer commands (#182)
* Add testing infrastructure

This adds testing infrastructure to test against multiple versions of
Rails, including Rails `main` branch.

* Test shared behavior of installer commands

This adds test coverage for the shared behavior of the installer
commands.  The commands are tested against a freshly generated Rails app
using the version of Rails that is currently loaded.  Thus the
installers can be tested with different versions of Rails in CI.

* Test specific behavior of javascript:install:bun

This adds test coverage for specific behavior of the
`javascript:install:bun` command.

* Test specific behavior of javascript:install:esbuild

This adds test coverage for specific behavior of the
`javascript:install:esbuild` command.

* Test specific behavior of javascript:install:rollup

This adds test coverage for specific behavior of the
`javascript:install:rollup` command.

* Test specific behavior of javascript:install:webpack

This adds test coverage for specific behavior of the
`javascript:install:webpack` command.